// apps/worker/src/types/onebot11/api.ts
export enum Action {
  /** 发送私聊消息 */
  sendPrivateMsg = 'send_private_msg',
  /** 发送群消息 */
  sendGroupMsg = 'send_group_msg',
  /** 发送消息 */
  sendMsg = 'send_msg',
  /** 撤回消息 */
  deleteMsg = 'delete_msg',
  /** 获取消息 */
  getMsg = 'get_msg',
  /** 获取转发消息 */
  getForwardMsg = 'get_forward_msg',
  /** 发送好友赞 */
  sendLike = 'send_like',
  /** 群组踢人 */
  setGroupKick = 'set_group_kick',
  /** 群组禁言 */
  setGroupBan = 'set_group_ban',
  /** 群组匿名用户禁言 */
  setGroupAnonymousBan = 'set_group_anonymous_ban',
  /** 群组全员禁言 */
  setGroupWholeBan = 'set_group_whole_ban',
  /** 设置群管理员 */
  setGroupAdmin = 'set_group_admin',
  /** 设置群匿名聊天 */
  setGroupAnonymous = 'set_group_anonymous',
  /** 设置群名片（群备注） */
  setGroupCard = 'set_group_card',
  /** 设置群名 */
  setGroupName = 'set_group_name',
  /** 退出群组 */
  setGroupLeave = 'set_group_leave',
  /** 设置群成员专属头衔 */
  setGroupSpecialTitle = 'set_group_special_title',
  /** 处理好友添加请求 */
  setFriendAddRequest = 'set_friend_add_request',
  /** 处理群添加请求／邀请 */
  setGroupAddRequest = 'set_group_add_request',
  /** 获取登录号信息 */
  getLoginInfo = 'get_login_info',
  /** 获取陌生人信息 */
  getStrangerInfo = 'get_stranger_info',
  /** 获取好友列表 */
  getFriendList = 'get_friend_list',
  /** 获取群信息 */
  getGroupInfo = 'get_group_info',
  /** 获取群列表 */
  getGroupList = 'get_group_list',
  /** 获取群成员信息 */
  getGroupMemberInfo = 'get_group_member_info',
  /** 获取群成员列表 */
  getGroupMemberList = 'get_group_member_list',
  /** 获取群荣誉信息 */
  getGroupHonorInfo = 'get_group_honor_info',
  /** 获取 Cookies */
  getCookies = 'get_cookies',
  /** 获取 CSRF Token */
  getCsrfToken = 'get_csrf_token',
  /** 获取 QQ 相关接口凭证 */
  getCredentials = 'get_credentials',
  /** 获取语音 */
  getRecord = 'get_record',
  /** 获取图片 */
  getImage = 'get_image',
  /** 是否可以发送图片 */
  canSendImage = 'can_send_image',
  /** 是否可以发送语音 */
  canSendRecord = 'can_send_record',
  /** 获取插件运行状态 */
  getStatus = 'get_status',
  /** 获取版本信息 */
  getVersionInfo = 'get_version_info',
  /** 获取版本信息 */
  getVersion = 'get_version',
  /** 重启插件 */
  setRestart = 'set_restart',
  /** 清理数据缓存 */
  cleanCache = 'clean_cache',
  /** 发送合并转发消息 */
  sendForwardMsg = 'send_forward_msg',
  /** 发送群聊合并转发消息 */
  sendGroupForwardMsg = 'send_group_forward_msg',
  /** 发送好友合并转发消息 */
  sendPrivateForwardMsg = 'send_private_forward_msg',
  /** 获取好友历史消息记录 */
  getFriendMsgHistory = 'get_friend_msg_history',
  /** 获取群组历史消息记录 */
  getGroupMsgHistory = 'get_group_msg_history',
  /** 对消息进行表情回应 */
  setMsgEmojiLike = 'set_msg_emoji_like',
  /** 上传群文件 */
  uploadGroupFile = 'upload_group_file',
  /** 上传私聊文件 */
  uploadPrivateFile = 'upload_private_file',
  /** 获取精华消息列表 */
  getEssenceMsgList = 'get_essence_msg_list',
  /** 设置精华消息 */
  setEssenceMsg = 'set_essence_msg',
  /** 删除精华消息 */
  deleteEssenceMsg = 'delete_essence_msg'
}
// ...
export interface GetGroupInfo {
  /** 群号 */
  group_id: number;
  /** 群名称 */
  group_name: string;
  /** 成员数 */
  member_count: number;
  /** 最大成员数（群容量） */
  max_member_count: number;
}

export interface HonorInfoList {
  /** QQ 号 */
  user_id: number;
  /** QQ 昵称 */
  nickname: string;
  /** 头像url */
  avatar: string;
  /** 荣誉描述 */
  description: string;
}

export interface GetGroupMemberInfo {
  /** 群号 */
  group_id: number;
  /** QQ 号 */
  user_id: number;
  /** 昵称 */
  nickname: string;
  /** 群名片／备注 */
  card: string;
  /** 性别 */
  sex: 'male' | 'female' | 'unknown';
  /** 年龄 */
  age: number;
  /** 地区 */
  area: string;
  /** 加群时间戳 */
  join_time: number;
  /** 最后发言时间戳 */
  last_sent_time: number;
  /** 成员等级 */
  level: string;
  /** 角色 */
  role: 'owner' | 'admin' | 'member' | 'unknown';
  /** 是否不良记录成员 */
  unfriendly: boolean;
  /** 专属头衔 */
  title: string;
  /** 专属头衔过期时间戳 */
  title_expire_time: number;
  /** 是否允许修改群名片 */
  card_changeable: boolean;
}